# swagger.yaml
swagger: '2.0'
info:
  title: gas
  version: 1.0.0
host: 'vvdev.ru:1334'
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
  /regions:
    get:
        tags:
        - Region
        responses:
          200:
            description: ok
            schema:
              type: array
              items:
                $ref: '#/definitions/Region'
    post:
      tags:
      - Region
      parameters:
      - in: body
        name: body
        schema:
          type: object
          $ref: '#/definitions/Region'
      responses:
        200:
          description: ok
          schema:
            $ref: '#/definitions/Region'
  /regions/{id}:
    get:
        tags:
        - Region
        parameters:
        - in: path
          name: id
          type: string
          required: true
        responses:
          200:
            description: ok
            schema:
              $ref: '#/definitions/Region'
    put:
      tags:
      - Region
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: body
        name: body
        schema:
          type: object
          $ref: '#/definitions/Region'
      responses:
        200:
          description: ok
          schema:
            $ref: '#/definitions/Region'
  /regions/{id}/logo:
    post:
      tags:
      - Region
      consumes: 
        - multipart/form-data
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: formData
        name: file
        required: true
        type: file
      responses:
        200:
          description: Название файла 
          schema:
            type: object
            properties:
              file:
                type: string
        500:
          description: Ошибка сервера
...
  /orders:
    post:
      tags:
      - Order
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            amount:
              type: number
            description:
              type: string
            addParams:
              type: object
              properties:
                string:
                  type: string
            approveUrl:
              type: string
            declineUrl:
              type: string
            cancelUrl:
              type: string
            email: 
              type: string
            contr_UID: 
              type: string
            cons_UID:
              type: string
      responses:
        200:
          description: ok
    get:
      tags:
      - Order
      responses:
        200:
          description: ok
  /orders/{id}:
    get:
      tags:
      - Order
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
  /orders/{id}/status:
    get:
      tags:
      - Order
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
  /orders/order_schedules:
    post:
      tags:
      - Order
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            amount:
              type: number
            description:
              type: string
            addParams:
              type: object
              properties:
                string:
                  type: string
            approveUrl:
              type: string
            declineUrl:
              type: string
            cancelUrl:
              type: string
            startDate:
              type: string
            repeatPeriodHours:
              type: number
            email: 
              type: string
            contr_UID: 
              type: string
      responses:
        200:
          description: ok
    get:
      tags:
      - Order
      responses:
        200:
          description: ok
  /orders/order_schedules/{id}:
    put:
      tags:
      - Order
      parameters:
      - in: path
        name: id
        required: true
        type: string
      - in: body
        name: body
        schema:
          type: object
          properties:
            amount:
              type: number
            paymentPerson:
              type: string
            inn:
              type: string
            kpp:
              type: string
            addParams:
              type: object
              properties:
                string:
                  type: string
            approveUrl:
              type: string
            declineUrl:
              type: string
            cancelUrl:
              type: string
            startDate:
              type: string
            repeatPeriodHours:
              type: number
      responses:
        200:
          description: ok
    get:
      tags:
      - Order
      parameters:
      - in: path
        name: id
        required: true
        type: string
      responses:
        200:
          description: ok
    delete:
      tags:
      - Order
      parameters:
      - in: path
        name: id
        required: true
        type: string
      responses:
        200:
          description: ok
  /super_admins/login:
    post:
      tags:
      - SuperAdmin
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
...
  /news:
    post:
      tags:
      - News
      summary: Создать новость
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            title:
              type: string
            body:
              type: string
            content:
              type: string
      responses:
        200:
          description: ok
    get:
      tags:
      - News
      summary: список новостей
      responses:
        200:
          description: ok
  /news/{id}:
    get:
      tags:
      - News
      summary: новость по id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
    put:
      tags:
      - News
      summary: обновить новость по id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: body
        name: body
        schema:
          type: object
          properties:
            title:
              type: string
            body:
              type: string
            content:
              type: string
      responses:
        200:
          description: ok
    delete:
      tags:
      - News
      summary: удалить новость по id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
...
  /super_admins/ssd_url:
    post:
      tags:
      - SSD
      summary: Создать адрес ссд
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            uri:
              type: string
            email:
              type: string
            title:
              type: string
      responses:
        200:
          description: Токен
    get:
      tags:
      - SSD
      summary: Список ссд
      responses:
        200:
          description: Токен 
  /super_admins/ssd_url/{id}:
    put:
      tags:
      - SSD
      summary: Обновить адрес ссд
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: body
        name: body
        schema:
          type: object
          properties:
            uri:
              type: string
            email:
              type: string
            title:
              type: string
      responses:
        200:
          description: Токен
    delete:
      tags:
      - SSD
      summary: удалить адрес ссд
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Токен
  /admins/login:
    post:
      tags:
      - Admin
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
...
definitions:
  Region: 
    type: object
    properties:
      _id:
        type: string
      logo: 
        type: string
        description: Название файла логотипа, можно получить по /media/:filename
      name: 
        type: string
        description: Имя региона
      pay: 
        type: object
        properties:
          typePaymentCode:
            type: string
            description: Вид перевода. Идентификатор, обозначающий, что платеж поступил из ЛК ЮЛ ЮЗДО, не нужно отправлять пока что
          lspu:
            type: string
            description: Лицевой счет у Поставщика услуг
          pointCode:
            type: string
            description: Точка приема
          abonentId:
            type: string
            description: Абонент
          idpu:
            type: string
            description: Поставщик
          service:
            type: string
            description: Услуга
      contacts:
        type: object
        properties:
          consumerAdministrationEmail: 
            type: string
            description: Управление по работе с потребителями газа
          communityAdministrationEmail: 
            type: string
            description: Управление по работе с населением
          supportEmail: 
            type: string
            description: Почта поддержки
          technicalSupportEmail: 
            type: string
            description: Почта технической поддержки
...